test(gyro_odometer): inline the node test helper into the test file
The three factory functions had their own header and source file but were used
by one translation unit only. Move them into an anonymous namespace in the test,
as autoware_vehicle_velocity_converter does, and delete the helper files.
