+++ localization/autoware_gyro_odometer/CMakeLists.txt
@@ -30,7 +30,6 @@ if(BUILD_TESTING)
     ament_add_ros_isolated_gtest(test_gyro_odometer
       test/test_main.cpp
       test/test_gyro_odometer_pubsub.cpp
-      test/test_gyro_odometer_helper.cpp
     )
     ament_target_dependencies(test_gyro_odometer
       rclcpp
